Changes Double to BigDecimal, still need to work with Frete

+++ src/main/java/br/unitins/topicos1/lgc/Frete/service/FreteService.java
@@ -1,10 +1,12 @@
 package br.unitins.topicos1.lgc.Frete.service;
 
+import java.math.BigDecimal;
+
 import br.unitins.topicos1.lgc.Endereco.model.Endereco;
 
 public interface FreteService {
     
     // Calcula o frete baseando-se no CEP (Estado)
-    Double calcularFrete(Endereco enderecoEntrega);
+    BigDecimal calcularFrete(Endereco enderecoEntrega);
     
 }
